[chore] Rename all .jsx files to use camelCase with the name matching the related React class name
