+++ client/app/components/App.jsx
@@ -2,7 +2,7 @@ import { h, Component } from 'preact'
 
 import { I18n } from '../plugins/preact-polyglot'
 
-import Sidebar from './sidebar'
+import Sidebar from './Sidebar'
 
 const App = ({ context, lang, categories, children }) => (
     <I18n context={context} lang={lang}>
