+++ client/app/components/CategoryList.jsx
@@ -1,11 +1,11 @@
 import { h, Component } from 'preact'
 import { translate } from '../plugins/preact-polyglot'
-import AccountList from './account_list'
+import AccountsList from './AccountsList'
 
 const CategoryList = ({ t, accounts, children }) => (
     <div class="content">
         <h1>{t('my_accounts category title')}</h1>
-        <AccountList accounts={filterUnconnected(accounts)} />
+        <AccountsList accounts={filterUnconnected(accounts)} />
         {children}
     </div>
 )
