+++ client/app/components/ConnectedList.jsx
@@ -1,11 +1,11 @@
 import { h, render } from 'preact'
 import { translate } from '../plugins/preact-polyglot'
-import AccountList from './account_list'
+import AccountsList from './AccountsList'
 
 const ConnectedList = ({ t, accounts, children }) => (
     <div class="content">
         <h1>{t('my_accounts connected title')}</h1>
-        <AccountList accounts={filterConnected(accounts)} />
+        <AccountsList accounts={filterConnected(accounts)} />
         {children}
     </div>
 )
